refactor(server): extract hosted collection service
